+++ frontend/dwaves-app/src/components/ContentAlbum.tsx
@@ -14,6 +14,8 @@ interface Props {
   setIsPlaying: React.Dispatch<React.SetStateAction<boolean>>
   setArtist : React.Dispatch<React.SetStateAction<AlbumDetail|undefined>>
   setAlert: React.Dispatch<React.SetStateAction<responseRequest | undefined>>
+  likedMusics: string[];
+  likeOrDislikeMusic: (music: string)=> void
 }
 
 export const ContentAlbum: React.FC<Props> = ({
@@ -23,7 +25,9 @@ export const ContentAlbum: React.FC<Props> = ({
   isPlaying,
   setIsPlaying,
   setArtist,
-  setAlert
+  setAlert,
+  likedMusics,
+  likeOrDislikeMusic
 }) => {
   const { id } = useParams()
 
@@ -75,6 +79,8 @@ export const ContentAlbum: React.FC<Props> = ({
         setIsPlaying={setIsPlaying} 
         setArtist={setArtist}
         setAlert={setAlert}
+        likedMusics={likedMusics}
+        likeOrDislikeMusic={likeOrDislikeMusic}
       />
     </div>
   )
