+++ frontend/dwaves-app/src/components/ExploPlayer.tsx
@@ -2,9 +2,10 @@ import "styles/Explorer.scss";
 import { Icon } from "components/shared";
 import styles from "styles/global/styles.module.scss";
 
-import { useRef } from "react";
+import React, {useEffect, useRef, useState} from "react";
 import { PlayPause, PlayRandomSong } from "songs/listenMusic";
 import { AlbumDetail, Music } from "models";
+import axios from "axios";
 
 interface Props {
   audioElmt: React.RefObject<HTMLAudioElement>;
@@ -20,6 +21,8 @@ interface Props {
   setRandom: React.Dispatch<React.SetStateAction<boolean>>;
   random: boolean;
   PlayRandomSong: (songs: Music[]) => Music;
+  likedMusics: string[];
+  likeOrDislikeMusic: (music: string)=> void
 }
 
 export const ExploPlayer: React.FC<Props> = ({
@@ -35,8 +38,20 @@ export const ExploPlayer: React.FC<Props> = ({
   repeat,
   setRandom,
   random,
-  PlayRandomSong
+  PlayRandomSong,
+  likedMusics,
+  likeOrDislikeMusic
 }) => {
+
+  const foundCidMusic = (musicUrl: string) => {
+    const musicCIDArray = musicUrl.split("/")
+    return musicCIDArray[musicCIDArray.length - 1]
+  }
+
+  const isLikedMusics = (music: string) => {
+    const musicCID = foundCidMusic(music)
+    return likedMusics.includes(musicCID);
+  }
 
   let index = songs.findIndex(
     (x: { name: string }) => x.name == currentSong.name
@@ -120,8 +135,13 @@ export const ExploPlayer: React.FC<Props> = ({
           songs &&
           <div id="contain-right-bar" className="flex row nowrap">
             <p>01:21 / 02:03</p>
-            <Icon icon="dislike" />
-            <Icon icon="close" />
+            <button className="pb-5" onClick={() => likeOrDislikeMusic(currentSong.src!)}>
+              <Icon icon="like" color={isLikedMusics(currentSong.src!) ? "red" : "black"}
+                    variant={isLikedMusics(currentSong.src!) ? "Bold" : "Linear"}/>
+            </button>
+            <div className="pt-1.5 pr-5">
+                <Icon icon="close" />
+            </div>
           </div>
         }
       </div>
